+++ core/src/forest/ForestTrainers.h
@@ -41,6 +41,19 @@ public:
   static ForestTrainer custom_trainer(Data* data,
                                       size_t outcome_index,
                                       double alpha);
+
+  static ForestTrainer regularized_regression_trainer(Data* data,
+                                                      size_t outcome_index,
+                                                      double lambda,
+                                                      bool downweight_penalty);
+
+  static ForestTrainer regularized_instrumental_trainer(Data* data,
+                                                        size_t outcome_index,
+                                                        size_t treatment_index,
+                                                        size_t instrument_index,
+                                                        double split_regularization,
+                                                        double lambda,
+                                                        bool downweight_penalty);
 };