+++ r-package/grf/R/RcppExports.R
@@ -21,8 +21,8 @@ custom_predict_oob <- function(forest_object, input_data, sparse_data, variable_
     .Call('grf_custom_predict_oob', PACKAGE = 'grf', forest_object, input_data, sparse_data, variable_names, num_threads)
 }
 
-instrumental_train <- function(input_data, outcome_index, treatment_index, instrument_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, split_regularization, alpha) {
-    .Call('grf_instrumental_train', PACKAGE = 'grf', input_data, outcome_index, treatment_index, instrument_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, split_regularization, alpha)
+instrumental_train <- function(input_data, outcome_index, treatment_index, instrument_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, split_regularization, alpha, lambda, downweight_penalty) {
+    .Call('grf_instrumental_train', PACKAGE = 'grf', input_data, outcome_index, treatment_index, instrument_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, split_regularization, alpha, lambda, downweight_penalty)
 }
 
 instrumental_predict <- function(forest_object, input_data, sparse_data, variable_names, num_threads, ci_group_size) {
@@ -45,8 +45,8 @@ quantile_predict_oob <- function(forest_object, quantiles, input_data, sparse_da
     .Call('grf_quantile_predict_oob', PACKAGE = 'grf', forest_object, quantiles, input_data, sparse_data, variable_names, num_threads)
 }
 
-regression_train <- function(input_data, outcome_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, alpha) {
-    .Call('grf_regression_train', PACKAGE = 'grf', input_data, outcome_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, alpha)
+regression_train <- function(input_data, outcome_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, alpha, lambda, downweight_penalty) {
+    .Call('grf_regression_train', PACKAGE = 'grf', input_data, outcome_index, sparse_data, variable_names, mtry, num_trees, verbose, num_threads, min_node_size, sample_with_replacement, keep_inbag, sample_fraction, no_split_variables, seed, honesty, ci_group_size, alpha, lambda, downweight_penalty)
 }
 
 regression_predict <- function(forest_object, input_data, sparse_data, variable_names, num_threads, ci_group_size) {