+++ r-package/grf/bindings/InstrumentalForestBindings.cpp
@@ -29,15 +29,25 @@ Rcpp::List instrumental_train(Rcpp::NumericMatrix input_data,
                               bool honesty,
                               unsigned int ci_group_size,
                               double split_regularization,
-                              double alpha) {
+                              double alpha,
+                              double lambda,
+                              bool downweight_penalty) {
   Data* data = RcppUtilities::convert_data(input_data, sparse_data, variable_names);
 
-  ForestTrainer trainer = ForestTrainers::instrumental_trainer(data,
-          outcome_index - 1,
-          treatment_index - 1,
-          instrument_index - 1,
-          split_regularization,
-          alpha);
+  ForestTrainer trainer = lambda > 0
+    ? ForestTrainers::regularized_instrumental_trainer(data,
+                                                       outcome_index - 1,
+                                                       treatment_index - 1,
+                                                       instrument_index - 1,
+                                                       split_regularization,
+                                                       lambda,
+                                                       downweight_penalty)
+    : ForestTrainers::instrumental_trainer(data,
+                                           outcome_index - 1,
+                                           treatment_index - 1,
+                                           instrument_index - 1,
+                                           split_regularization,
+                                           alpha);
   RcppUtilities::initialize_trainer(trainer, mtry, num_trees, num_threads, min_node_size,
       sample_with_replacement, sample_fraction, no_split_variables, seed, honesty, ci_group_size);