+++ releases/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.2] - 2017-07-06
+
+### Added
+- Add an experimental regularized version of the regression splitting rule.
+
+### Fixed
+- Several bugfixes for CRAN compatibility.
+
 ## [0.9.1] - 2017-06-27
 
 ### Added